app-emulation/runc: bump 1.1.5
This is the fifth patch release in the 1.1.z series of runc, which fixes
three CVEs found in runc.

    * CVE-2023-25809 is a vulnerability involving rootless containers where
      (under specific configurations), the container would have write access
      to the /sys/fs/cgroup/user.slice/... cgroup hierarchy. No other
      hierarchies on the host were affected. This vulnerability was
      discovered by Akihiro Suda.
      https://github.com/opencontainers/runc/security/advisories/GHSA-m8cg-xc2p-r3fc

    * CVE-2023-27561 was a regression which effectively re-introduced
      CVE-2019-19921. This bug was present from v1.0.0-rc95 to v1.1.4. This
      regression was discovered by @Beuc.
      GHSA-vpvm-3wq2-2wvm

    * CVE-2023-28642 is a variant of the same bug and was fixed by the same
      patch. This variant of the above vulnerability was reported by Lei
      Wang.
      https://github.com/opencontainers/runc/security/advisories/GHSA-g2j6-57v7-gm8c

In addition, the following other fixes are included in this release:

    * Fix the inability to use /dev/null when inside a container. (#3620)
    * Fix changing the ownership of host's /dev/null caused by fd redirection
      (a regression in 1.1.1). (#3674, #3731)
    * Fix rare runc exec/enter unshare error on older kernels, including
      CentOS < 7.7. (#3776)
    * nsexec: Check for errors in write_log(). (#3721)

Static Linking Notices

The runc binary distributed with this release are statically linked with
the following GNU LGPL-2.1 licensed libraries, with runc acting
as a "work that uses the Library":

    libseccomp

The versions of these libraries were not modified from their upstream versions,
but in order to comply with the LGPL-2.1 (§6(a)), we have attached the
complete source code for those libraries which (when combined with the attached
runc source code) may be used to exercise your rights under the LGPL-2.1.

However we strongly suggest that you make use of your distribution's packages
or download them from the authoritative upstream sources, especially since
these libraries are related to the security of your containers.